[Render engine] Rework and simplify part 2

+++ razarion-client/src/main/java/com/btxtech/client/editor/renderer/MonitorRenderTask.java
@@ -1,9 +1,9 @@
 package com.btxtech.client.editor.renderer;
 
 import com.btxtech.uiservice.renderer.AbstractRenderTask;
-import com.btxtech.uiservice.renderer.CommonRenderComposite;
 import com.btxtech.uiservice.renderer.ModelRenderer;
 import com.btxtech.uiservice.renderer.RenderService;
+import com.btxtech.uiservice.renderer.RenderSubTask;
 import com.btxtech.uiservice.renderer.RenderUnitControl;
 import com.google.inject.Inject;
 
@@ -15,28 +15,34 @@ import javax.enterprise.context.ApplicationScoped;
  */
 @ApplicationScoped
 public class MonitorRenderTask extends AbstractRenderTask<Void> {
+    public interface SubTask extends RenderSubTask<Void> {
+    }
+
     @Inject
     private RenderService renderService;
     private boolean showDeep = true;
+    private ModelRenderer<Void> modelRenderer;
 
     public void showMonitor() {
-        ModelRenderer<Void> modelRenderer = create();
-        CommonRenderComposite<ShadowMonitorRendererUnit, Void> renderComposite = modelRenderer.create();
-        renderComposite.init(null);
-        renderComposite.setRenderUnit(ShadowMonitorRendererUnit.class);
-        modelRenderer.add(RenderUnitControl.NORMAL, renderComposite);
-        add(modelRenderer);
-        renderComposite.fillBuffers();
+        if(modelRenderer != null) {
+            throw new IllegalStateException("Shadow monitor is already showing");
+        }
+        modelRenderer = createNew();
+        modelRenderer.create(RenderUnitControl.NORMAL, MonitorRenderTask.SubTask.class, null);
         renderService.addRenderTask(this, "Monitor");
     }
 
     public void hideMonitor() {
+        if(modelRenderer == null) {
+            throw new IllegalStateException("Shadow monitor is already showing");
+        }
         renderService.removeRenderTask(this);
-        clear();
+        destroy(modelRenderer);
+        modelRenderer = null;
     }
 
     public boolean isShown() {
-        return renderService.containsRenderTask(this);
+        return modelRenderer != null;
     }
